[M] Refactor Code Base [M] Rename Json Node [A] Rest Template, Service to fetch data from third party api [A] Add Open Weather configuration
[M] Move package client
[M] longitude, latitude field in City
[A] last update field to Historical Response
[A] convert utc time to local time

[M] insert.sql script to insert city only null

[M] Fix Test Case does not compile
[M] Remove Date API from Java 8
[A] Add Qualifier for bean

[M] Fix Test Case does not compile

[M] Fix LocalDate pattern in param
[A] Exception Handler

[M] Fix Weather can not be added

+++ weather/src/main/resources/sql/create.sql
@@ -2,8 +2,8 @@ create table city
 (
     id          varchar(36) not null unique
         primary key,
-    latitude    bigint       not null,
-    longitude   bigint       not null,
+    latitude    double        not null,
+    longitude   double       not null,
     name        varchar(255) null,
     create_time datetime     null default (UTC_TIMESTAMP()),
     update_time datetime     null default (UTC_TIMESTAMP())
